% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clinical_trials.R
\name{filter_clinical_trials}
\alias{filter_clinical_trials}
\title{Filter a set of PMIDs for clinical trials}
\usage{
filter_clinical_trials(pmid, phase = 0, ...)
}
\arguments{
\item{pmid}{(vec, char, or list) Either raw output from \link{query_pubmed} or
one or more PMIDs.}

\item{phase}{(int; can be , 1, 2, 3, or 4; default 0) FDA-style clinical trial phase to filter for.
Enter 0 or keep blank to include all clinical trials.}

\item{...}{Arguments to pass to \link{filter_pub_types}.}
}
\value{
Returns a list object in the same format as \link{query_pubmed}.
}
\description{
This function allows the user to filter an already acquired set of PMIDs for those that
are associated with a clinical trial, as well as allowing the user to select for FDA-
style clinical trials by phase.
}
