% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mesh_terms.R
\name{filter_mesh_terms}
\alias{filter_mesh_terms}
\title{Filter a set of PMIDs for articles that include a set of MeSH terms}
\usage{
filter_mesh_terms(pmid, mesh_term, ...)
}
\arguments{
\item{pmid}{(vec, char, or list) Either raw output from \link{query_pubmed} or
one or more PMIDs.}

\item{mesh_term}{(vec or char) One or more MeSH terms that you would like to include in your results.}

\item{...}{Arguments to pass to \link{query_pubmed}.}
}
\value{
Returns a list object in the same format as \link{query_pubmed}.
}
\description{
This function allows the user to filter an already acquired set of PMIDs for those that
are associated with one or more specific MeSH terms (e.g., "Clinical Trials, Phase I as Topic")
}
