% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pub_types.R
\name{filter_pub_types}
\alias{filter_pub_types}
\title{Filter a set of PMIDs for articles that match a specific publication type}
\usage{
filter_pub_types(pmid, pub_type, ...)
}
\arguments{
\item{pmid}{(vec, char, or list) Either raw output from \link{query_pubmed} or
one or more PMIDs.}

\item{pub_type}{(vec or char) One or more MeSH terms that you would like to include in your results.}

\item{...}{Arguments to pass to \link{query_pubmed}.}
}
\value{
Returns a list object in the same format as \link{query_pubmed}.
}
\description{
This function allows the user to filter an already acquired set of PMIDs for those that
match a specific publication type (e.g., "Clinical Trial, Phase I").
}
