% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citing_articles.R
\name{get_citing_articles}
\alias{get_citing_articles}
\title{Query PubMed}
\usage{
get_citing_articles(pmid)
}
\arguments{
\item{pmid}{(char, vec, or list) Either raw output from \link{query_pubmed}, or a vector of one or more
PMIDs that you want to find citations for, or a single PMID.}
}
\value{
A vector of PMIDs corresponding to articles that cite the provided PMID(s).
}
\description{
This function uses \code{rentrez::entrez_link}
function to find citations of specified article(s). See \code{rentrez}
documentation for more details.
}
