% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nsf_api.R
\name{query_nsf_api}
\alias{query_nsf_api}
\title{Query NSF API}
\usage{
query_nsf_api(
  printFields = "standard",
  export_path = "",
  export_fmt = "csv",
  r_output = TRUE,
  help = FALSE,
  rpp = 25,
  ...
)
}
\arguments{
\item{printFields}{(default: "standard") Fields to print. If "standard", the API's default print
fields will be included. If "all", every possible print field will be included. Otherwise,
you can use a vector with the specific fields you wish to include. See link from \code{help}
for more information.}

\item{export_path}{(default: "") Allows you to write the output to the specified file
path.}

\item{export_fmt}{(default: "csv") Allows you to choose file format for \code{export_path}.
Options are "csv", "excel", "rdata", "spss", and "sas". (NOTE: if using the "csv" setting,
\code{printFields} is likely to cause an error if set to "all"; use a different file type
or narrow your field selection in this case.)}

\item{r_output}{(default: TRUE) Allows you to choose whether or not to generate any output in
R. Keep as TRUE if you plan to use the output in this R session; switch to FALSE if you
only want to generate an export file.}

\item{help}{(default: FALSE) If TRUE, opens https://resources.research.gov/common/webapi/awardapisearch-v1.htm#request-parameters-notes
which contains information about API fields.}

\item{rpp}{(default: 25) Results per page; this gets passed to the API query. Set to the API
maximum by default.}

\item{...}{Additional parameters to be included in the API query. You can find a list at the
.   webpage referenced when using the \code{help} parameter. Use either a string or a vector.
.   If vector, elements will be bound with "+", corresponding to an "OR" operator. If string,
it will be searched in double quotations. Note that vectors CANNOT contain multiple
multi-word terms; if you want to search both "University of Iowa" and "Iowa State University",
for instance, this will need to be two different queries.}
}
\value{
A data frame containing information matching the criteria provided.
}
\description{
This function is a wrapper for the NSF Awards API. It allows the user to input the API parameters
into the function as regular R parameters.
}
\examples{
library(fedscifund)

# query the API for Iowa grants containing the keyword "neuroscience" for all fields
df <- query_nsf_api(
  printFields = "all",
  perfStateCode = "IA",
  keyword = "neuroscience"
  ) 

# run the same query as above but only retrieve title and abstract
df <- query_nsf_api(
  printFields = c("title", "abstractText"), 
  perfStateCode = "IA", 
  keyword = "neuroscience"
  ) 

# run the above code but export results to Excel without creating an R object
query_nsf_api(
  printFields = c("title", "abstractText"), 
  export_path = "ia_neuro.xlsx", 
  export_fmt = "excel", 
  r_output = FALSE, 
  perfStateCode = "IA", 
  keyword = "neuroscience"
  )
}
