% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_pubmed.R
\name{query_pubmed}
\alias{query_pubmed}
\title{Query PubMed}
\usage{
query_pubmed(query, retmax = 9999, unlist = TRUE, browser = FALSE, ...)
}
\arguments{
\item{query}{(char) This is your PubMed query. Outputs from \code{create_pubmed_query}
can go here.}

\item{retmax}{(int) Maximum number of records to be returned, set by default to the max that
the PubMed API will allow.}

\item{unlist}{(logical, default = FALSE) Opens a PubMed browser session with the entered query.}

\item{...}{Additional arguments accepted by the \code{entrez_search} function, should you wish
to include them.}
}
\value{
A \code{list} object with information about the results (including PMIDs).
}
\description{
This function uses the \code{rentrez::entrez_search}
function to search PubMed. See \code{rentrez} documentation for more details. The only
significant change is that
}
